Move recognition of digit "one" and debug output with digit coordinates below vertical segmentation.
Before this the recognition of a "one" used the image height instead of
the digit height. This kind of worked because for all practical purposes
the image was cropped to leave just the digits.

The debug output of digit coordinates was flawed the same way.

Since now the real height/width ratio is used to recognize a "one" the
empirically determined default value had to be adjusted as well. Using this
value my regression tests are all passed.