Avoid division by zero during recognition of "one" and "decimal point".
Do not divide by digit width or height if it is zero, just skip this
digit during special case recognition of a "one" or a "decimal point".

Currently, the scan line detection will see the following if digit width
or height is zero:
 _
|
|

This results in printing an underscore ('_') character for a not
recognized digit.

Bump version number to 2.14.1 for this bug fix.