ensure debug image shows bitmap used for recognition
The recognition algorithm always works on a bitmap of set and unset
(foreground and background) pixels. The debug image used to show the
input image after image processing, but if the processing did not include
a command that replaced pixels with fore- or background pixel values,
it would not show which pixels are foreground, and which are not. This
commit uses the make_mono() function to create the debug image, thereby
creating a representation of the bitmap used for recognition.