add option -S, --ascii-art-segments
This options prints the recognized segments, i.e. the display as
seen by ssocr, to standard error. This can be used as a quick check
what went wrong if the recognition does not work. Additionally,
it can be used to get the raw segment data to use a separate
program to interpret set segments as digits or characters.