move building a .deb file out of the way
This is intended to interfere less (not at all?) with Debian
packaging efforts.

The Debian Upstream Guide says:

   "Some projects include a rough /debian directory among source files
    to ease bleeding-edge package compilation and installation on Debian
    (and derived) systems. While this is a good effort, it is better to
    leave it out of the final tarball as it can interfere with debian's
    own packaging effort."

I have renamed the directory to "notdebian" to clearly show that this
is not official. Since the Debian packaging system needs the exact name
"debian" for this directory, "notdebian" does not interfere with this.