README: clarify that the official homepage is official
Many people, including packagers for GNU/Linux distributions, seem to
think that ssocr is a "GitHub project." It is not. It is older than
GitHub and may well exist longer that GitHub. It is older than git.

The official website contains documentation not found on GitHub. GitHub
is just use as a currently convenient way to have an external copy, a
backup, of the development repository. This does provide third parties
with access to the developent code and its history as an added benefit.

The "added value" of GitHub (issues, pull requests, ...) is meaningless
to me and may be abandoned anytime. The contents of the official website
will be preserved even if it needs to be moved. The development history
has already been preserved during the migration from subversion to git.