always use underscore '_' for unknown digits
Before, there was a special case where a blank ' ' could be printed
if no segments were detected in a digit at all. But this would not
happen, since all digits contain at least one pixel which would result
in some set segments.

This commit is a preparatory step for refactoring segement detection
in a digit.