refactor scanning for set segments
This introduces a function to scan part of the image for foreground
pixels.

This scanline() function may be of use for distuingishing between
the digit '1' and the symbol ":".

It may also help in segment detection reliability if the "len" parameter
is used to skip scanning image areas between segment positions.