defines.h: add commented out three segment N
While I know of seven segment displays that use
     _
    | |

as 'N', ssocr detects the size of the individual seven segment
displays and thus always sees this as

     _
    | |
    | |

unless the digit boundary detection fails.

This commit is intended as a reminder for me that I have already
thought about this issue.