guard against potential unsigned integer overflow
If the string length of the value of the environment variable TMP
is too big, adding the length of a slash and the temporary file
name might overflow, which would result in an insufficient memory
allocation for the absolute file name.

I do not know if that can actually happen in any existing operating
system and platform combination where ssocr can be used.