accept any number of digits during segmentation
If a specific number of digits is expected, i.e., without
-d-1, the number of found (potential) digits is compared
with the expected number, and ssocr errors out on mismatch.

This prepares for the introduction of an option to reject
some potential images, e.g., because they are too small to
contain a digit or character from the display.

It also prepares for allowing a range of digits as the
expected segmentation result, e.g., for a clock display
with a blinking colon (:), or for a thermometer display.