man page: fix "white_border" description
The command name stems from the time when ssocr always considered
"white" as background and "black" as foreground.  The intention
has always been to ensure some background border around the seven
segment displays, and the implementation always followed this.  The
output of --help correctly uses "background" for the "white_border"
description.  Now, the man page description is correct, too.